requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
python-socketio==5.10.0
deepl==1.16.1
httpx==0.25.2
alembic==1.13.1
bcrypt==4.1.2